+++ libexec/scoop-config.ps1
@@ -43,14 +43,14 @@
 #       * An empty or unset value for proxy is equivalent to 'default' (with no username or password)
 #       * To bypass the system proxy and connect directly, use 'none' (with no username or password)
 #
-# default-architecture: 64bit|32bit
+# default_architecture: 64bit|32bit
 #       Allow to configure preferred architecture for application installation.
 #       If not specified, architecture is determined be system.
 #
 # debug: $true|$false
 #       Additional and detailed output will be shown.
 #
-# force-update: $true|$false
+# force_update: $true|$false
 #       Force apps updating to bucket's version.
 #
 # show_update_log: $true|$false
@@ -80,6 +80,12 @@
 # virustotal_api_key:
 #       API key used for uploading/scanning files using virustotal.
 #       See: 'https://support.virustotal.com/hc/en-us/articles/115002088769-Please-give-me-an-API-key'
+#
+# cat_style:
+#       When set to a non-empty string, Scoop will use 'bat' to display the manifest for
+#       the `scoop cat` command and while doing manifest review. This requires 'bat' to be
+#       installed (run `scoop install bat` to install it), otherwise errors will be thrown.
+#       The accepted values are the same as ones passed to the --style flag of 'bat'.
 #
 # ignore_running_processes: $true|$false
 #       When set to $false (default), Scoop would stop its procedure immediately if it detects